voice-agent-websocket/requirements.txt
websocket-client==1.8.0
pyaudio
python-dotenv
